refactor(InputLabel): moved to internal
