+++ src/drafts/MarkdownEditor/Label.tsx
@@ -1,5 +1,5 @@
 import React, {FC, useContext} from 'react'
-import InputLabel from '../../_InputLabel'
+import InputLabel from '../../internal/components/InputLabel'
 import {SxProp} from '../../sx'
 import {MarkdownEditorContext} from './_MarkdownEditorContext'
