+++ src/internal/components/InputLabel.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import Box from './Box'
-import {SxProp} from './sx'
-import VisuallyHidden from './_VisuallyHidden'
+import Box from '../../Box'
+import {SxProp} from '../../sx'
+import VisuallyHidden from '../../_VisuallyHidden'
 
 type BaseProps = SxProp & {
   disabled?: boolean
